forecast_service/requirements.txt
argcomplete==1.9.2
arrow==0.12.1
attrs==17.4.0
base58==0.2.4
binaryornot==0.4.4
boto3==1.5.34
botocore==1.8.48
cachetools==2.0.1
certifi==2018.1.18
cfn-flip==1.0.0
chardet==3.0.4
click==6.7
colorama==0.3.9
dill==0.2.7.1
docutils==0.14
durationpy==0.5
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
future==0.16.0
futures==3.1.1
google-api-core==0.1.4
google-auth==1.4.1
google-cloud-core==0.28.0
google-cloud-firestore==0.28.0
google-cloud-pubsub==0.33.1
google-gax==0.15.16
googleapis-common-protos==1.5.3
grpcio==1.9.1
gunicorn==19.7.1
hjson==3.0.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jinja2-time==0.2.0
jmespath==0.9.3
kappa==0.6.0
MarkupSafe==1.0
numpy==1.14.0
pandas==0.22.0
placebo==0.8.1
pluggy==0.6.0
ply==3.8
poyo==0.4.1
protobuf==3.5.1
psycopg2==2.7.3.2
py==1.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyJWT==1.5.3
PySocks==1.6.8
pytest==3.4.0
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2018.3
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.13
six==1.11.0
SQLAlchemy==1.2.2
toml==0.9.4
tqdm==4.19.1
troposphere==2.2.0
Unidecode==1.0.22
urllib3==1.22
Werkzeug==0.12
whichcraft==0.4.1
wsgi-request-logger==0.4.6
